working on new word categorizer; reorganized classes
